Added support for theming the container that wraps WidgetBase instances, and also reverts plippard's changes(changeset 2636edfaf74c) of WidgetBase back to its previous version.
The previous WidgetBase's Render method rendered a lot of extra empty div container markup for the purpose of
theming. A good idea, but for most themes the extra markup was going unused.

Notes on this version:
1) WidgetBase no longer renders any extra container markup.
2) Theme creators now have more flexibility. They can just plop a WidgetContainer.ascx user control in their
theme directory and be all set to go. Users no longer have to customize and recompile WidgetBase in order to
edit this.
3) There's backwards compatiblity for all existing themes that do not contain a WidgetContainer. An internal
class, DefaultWidgetContainer, is used for rendering when a custom WidgetContainer can't be found. WidgetBase's
original rendering implementation has been moved to this class instead.
4) Added WidgetContainer.ascx to the Standard theme to show its basic implementation details.