Updated the Posts and Pages to have the aJax file manager Updated the Posts and pages attach buttons to use the new image handler Resolved bug with image cropping Resolved bug with image resizing Resolved bug with tinyMCE over writing content while using async uploadFMD.png Changed the get files method in the DbFileSystemProvider.cs to dispose of the data context immediatialy after calling image.png Updated the delete files to actually submit the delete.
Quick tested all the various methods.

TODO: Work out the GetDirectory() method to add an over load to auto create the directory. I have taken the CreateDirectory() and made it an
internal method as the CreateDirectory has alot of rules and in a datastore we could run into missing files.