Added a PreloadWidgetsAsync() method to decrease time required for a cold server to serve the first page.  Method called by the Global.asax.
Also made a small change to the OnLoad method by adding conditional logic to
only load the admin widget if the visitor is authenticated as an admin.
Revision control makes it look like more was done than really was.  It's
actually just one line of code to calculate the isAdminWidget bool
and then an if block to conditionally load the control if it is.
This approach decreased the time required for a cold server to serve the
first pay be approx 1 sec if the visitor isn't an admin. (for example the
google bot, indexing the site in the middle of the night)